Rename : getMetadata → getUrlmetadata
